// src/app/models/day-status.ts
export enum DayStatus {
    Work = 'work',
    Rest = 'rest',
    Annual = 'annual',
    HalfAnnual = 'half-annual',
    Casual = 'casual',
    HalfCasual = 'half-casual',
    Compensation = 'compensation',
    TwoHours = '2-hours',
    AbsentWithPermission = 'absent-with-permission',
    AbsentWithoutPermission = 'absent-without-permission',
    Sick = 'sick'
}